requirements.txt
fastapi
uvicorn
pydantic
requests
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv 